StockData/pyproject.toml
[project]
name = "stockdata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gm>=3.0.178",
    "pymongo>=4.14.1",
    "python-dotenv>=1.1.1",
    "schedule>=1.2.2",
    "pydantic>=2.0.0",
]
